Finished get controllers for shipment, added HasPrecision

+++ API/Model/LettersBag.cs
@@ -3,7 +3,7 @@ namespace API.Model
     public class LettersBag : Bag
     {
         public int LettersCount { get; set; }
-        public double Weight { get; set; }
-        public double Price { get; set; }
+        public decimal Weight { get; set; }
+        public decimal Price { get; set; }
     }
 }
